Fix node modules folder
