--- bookstore/src/Book/bookItem.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-const BookItem = ({ book }) => {
-    return (
-      <div>
-        <h1>{book.title}</h1>
-        <p>{book.author}</p>
-      </div>
-    );
-};
-
-export default BookItem;
