--- bookstore/src/Categories.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Categories = () => {
-  return (
-    <div>
-      <button type="button">check status</button>
-    </div>
-  );
-};
-
-export default Categories;
