+++ src/App.js
@@ -4,16 +4,16 @@ import BookContainer from './Book/bookcontainer';
 import Categories from './Categories';
 import Navbar from './Navbar';
 
-const App = () => {
-  return (
-    <>
+const App = () => (
+  <>
     <Navbar />
+    <div>
       <Routes>
         <Route path="/" element={<BookContainer />} />
         <Route path="/categories" element={<Categories />} />
       </Routes>
-    </>
-  );
-};
+    </div>
+  </>
+);
 
 export default App;
