+++ src/Book/bookItem.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+const BookItem = ({ book }) => (
+  <div>
+    <h1>{book.title}</h1>
+    <p>{book.author}</p>
+  </div>
+);
+
+export default BookItem;
