+++ src/Categories.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Categories = () => (
+  <div>
+    <button type="button">check status</button>
+  </div>
+);
+
+export default Categories;
