+++ src/Navbar.js
@@ -1,0 +1,16 @@
+import React from 'react';
+// import { Link } from 'react-router-dom';
+
+const Navbar = () => {
+  <div>
+    <nav className="navbar">
+      <h1>Book store</h1>
+      {/* <ul>
+        <li><Link to="/">Books</Link></li>
+        <li><Link to="/categories">Categories</Link></li>
+      </ul> */}
+    </nav>
+  </div>
+};
+
+export default Navbar;
